+++ src/Solidity by Example/variables.sol
@@ -1,12 +1,13 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
-contract Variables{
-    uint256 public num=123;
-    string public text="Hello";
-    function doSamething()public {
-    uint i=456;
-    uint256 timestamp=block.timestamp;
-    address sender=msg.sender;
+contract Variables {
+    uint256 public num = 123;
+    string public text = "Hello";
+
+    function doSamething() public {
+        uint256 i = 456;
+        uint256 timestamp = block.timestamp;
+        address sender = msg.sender;
     }
 }
